overflow counter as uint8_t
This reduced the size from 298 bytes to 282.  'int' involves a bunch of
instructions to deal with two byte numbers that uint8_t doesn't have to
deal with.

+++ c/avr/failsafe/failsafe.c
@@ -13,7 +13,7 @@
 #define BUTTON_OVERFLOWS 30
 
 volatile uint8_t prevb = 0xFF;
-volatile int overflows = 0;
+volatile uint8_t overflows = 0;
 
 // This part of the ISR(PCINT0_vect) runs when the button pin changes.
 static inline void isr_svc_button() {
